kubernetes_app/db_service/requirements.txt
fastapi
uvicorn
boto3
python-decouple
python-multipart